Catch server hang early in e2e tester

+++ protocol/test/e2e_helpers/dev_instance_tools.py
@@ -172,7 +172,11 @@ def apply_dpkg_locking_fixup(pexpect_process, pexpect_prompt, running_in_ci):
 
 
 def configure_aws_credentials(
-    pexpect_process, pexpect_prompt, running_in_ci, aws_credentials_filepath="~/.aws/credentials"
+    pexpect_process,
+    pexpect_prompt,
+    running_in_ci,
+    use_apt,
+    aws_credentials_filepath="~/.aws/credentials",
 ):
     """
     Configure AWS credentials on a remote machine by copying them from the ones configures on the machine where this script is being run.
@@ -232,20 +236,30 @@ def configure_aws_credentials(
     elif not running_in_ci:
         pexpect_process.expect(pexpect_prompt)
 
-    # Download and install AWS cli manually to avoid frequent apt install outages
-    # https://docs.aws.amazon.com/cli/latest/userguide/getting-started-install.html
-    pexpect_process.sendline(
-        "curl https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip -o awscliv2.zip"
-    )
-    wait_until_cmd_done(pexpect_process, pexpect_prompt, running_in_ci)
-    pexpect_process.sendline("sudo apt install -y unzip")
-    wait_until_cmd_done(pexpect_process, pexpect_prompt, running_in_ci)
-    pexpect_process.sendline("unzip awscliv2.zip")
-    wait_until_cmd_done(pexpect_process, pexpect_prompt, running_in_ci)
-    pexpect_process.sendline("rm awscliv2.zip")
-    wait_until_cmd_done(pexpect_process, pexpect_prompt, running_in_ci)
-    pexpect_process.sendline("sudo ./aws/install")
-    wait_until_cmd_done(pexpect_process, pexpect_prompt, running_in_ci)
+    if use_apt:
+        # Download and install AWS cli using apt-get
+        pexpect_process.sendline("sudo apt-get install awscli")
+        result = pexpect_process.expect(["Do you want to continue?", pexpect_prompt])
+        if result == 0:
+            pexpect_process.sendline("Y")
+            wait_until_cmd_done(pexpect_process, pexpect_prompt, running_in_ci)
+        elif not running_in_ci:
+            pexpect_process.expect(pexpect_prompt)
+    else:
+        # Download and install AWS cli manually to avoid frequent apt install outages
+        # https://docs.aws.amazon.com/cli/latest/userguide/getting-started-install.html
+        pexpect_process.sendline(
+            "curl https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip -o awscliv2.zip"
+        )
+        wait_until_cmd_done(pexpect_process, pexpect_prompt, running_in_ci)
+        pexpect_process.sendline("sudo apt install -y unzip")
+        wait_until_cmd_done(pexpect_process, pexpect_prompt, running_in_ci)
+        pexpect_process.sendline("unzip awscliv2.zip")
+        wait_until_cmd_done(pexpect_process, pexpect_prompt, running_in_ci)
+        pexpect_process.sendline("rm awscliv2.zip")
+        wait_until_cmd_done(pexpect_process, pexpect_prompt, running_in_ci)
+        pexpect_process.sendline("sudo ./aws/install")
+        wait_until_cmd_done(pexpect_process, pexpect_prompt, running_in_ci)
 
     pexpect_process.sendline("aws configure")
     pexpect_process.expect("AWS Access Key ID")
